+++ packages/admin/src/snackbar/UndoSnackbar.tsx
@@ -1,44 +1,38 @@
-import { Button, Slide, Snackbar } from "@material-ui/core";
+import { Button, Slide, Snackbar, SnackbarProps } from "@material-ui/core";
 import { TransitionProps } from "@material-ui/core/transitions";
 import * as React from "react";
 import { FormattedMessage } from "react-intl";
 
 import { useSnackbarApi } from "./SnackbarProvider";
 
-export interface SnackbarState<Payload extends unknown> {
+export interface UndoSnackbarProps<Payload extends unknown> extends Omit<SnackbarProps, "action"> {
     key: React.Key;
     message: React.ReactNode;
-    onUndoClick?: (payload?: Payload) => void;
+    onUndoClick: (payload?: Payload) => void;
     payload?: Payload;
 }
 
-export const useUndoSnackbar = () => {
+const UndoSnackbar = <Payload extends unknown>({ onUndoClick, payload, ...props }: UndoSnackbarProps<Payload>) => {
     const snackbarApi = useSnackbarApi();
 
-    const handleActionButtonClick = <Payload,>(snackbarState?: SnackbarState<Payload>) => {
+    const onClick = () => {
         snackbarApi.hideSnackbar();
-
-        snackbarState?.onUndoClick?.(snackbarState?.payload);
+        onUndoClick(payload);
     };
 
-    const updateSnackbarState = <Payload extends unknown>(newState: SnackbarState<Payload>) => {
-        snackbarApi.showSnackbar(
-            <Snackbar
-                anchorOrigin={{ vertical: "bottom", horizontal: "left" }}
-                key={newState?.key}
-                autoHideDuration={5000}
-                message={newState?.message}
-                action={
-                    <Button color="secondary" size="small" onClick={() => handleActionButtonClick(newState)}>
-                        <FormattedMessage id="cometAdmin.generic.undo" defaultMessage="Undo" />
-                    </Button>
-                }
-                TransitionComponent={(props: TransitionProps) => <Slide {...props} direction="right" />}
-            />,
-        );
-    };
-
-    return {
-        showUndoSnackbar: updateSnackbarState,
-    };
+    return (
+        <Snackbar
+            anchorOrigin={{ vertical: "bottom", horizontal: "left" }}
+            autoHideDuration={5000}
+            action={
+                <Button color="secondary" size="small" onClick={onClick}>
+                    <FormattedMessage id="cometAdmin.generic.undo" defaultMessage="Undo" />
+                </Button>
+            }
+            TransitionComponent={(props: TransitionProps) => <Slide {...props} direction="right" />}
+            {...props}
+        />
+    );
 };
+
+export default UndoSnackbar;